requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
oauthlib>=3.3.1
pyjwt>=2.10.1
flask-dance>=7.1.0
werkzeug>=3.1.3
stripe>=12.3.0
pandas>=2.3.0
sqlalchemy>=2.0.41
flask-limiter>=3.12
flask-wtf>=1.2.2
flask-mail>=0.10.0
pyotp>=2.9.0
qrcode>=8.2
python-dotenv>=1.0.0
azure-storage-blob>=12.0.0
azure-identity>=1.0.0